# TFM_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFM2 API Documentation",
        "description": "DocumentaciÃ³n de la API de TFM2",
        "contact": {
            "email": "tuemail@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "List all services",
                "operationId": "56a18686444b05a24ba02aa0d58c4a2d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServiceBasic"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service (auth required)",
                "operationId": "ed38e15a00a693a49e8dcae1ef352740",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "properties": {
                                            "street": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "zip": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "subcategory_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceBasic"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "price": [
                                                    "The price field must be a number."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServiceBasic": {
                "title": "ServiceBasic",
                "description": "Service fields without relations",
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Computer Repair"
                    },
                    "email": {
                        "type": "string",
                        "example": "robertosarta@gmail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34 600 987 654"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "456 Service St"
                            },
                            "city": {
                                "type": "string",
                                "example": "Barcelona"
                            },
                            "zip": {
                                "type": "string",
                                "example": "08001"
                            }
                        },
                        "type": "object"
                    },
                    "description": {
                        "type": "string",
                        "example": "Repair and maintenance of laptops and PCs"
                    },
                    "price": {
                        "type": "string",
                        "example": "49.99"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "subcategory_id": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Paste only the token; UI will add the Bearer prefix"
            }
        }
    },
...
}